rag-scrap/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
httpx>=0.27.0
supabase>=2.6.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
pypdf>=4.2.0
PyMuPDF>=1.24.3
pytesseract>=0.3.10
Pillow>=10.3.0
beautifulsoup4>=4.12.3
requests>=2.31.0
duckduckgo-search>=6.1.7
selenium>=4.21.0
webdriver-manager>=4.0.1
firecrawl>=1.4.0
smolagents>=1.2.0
PyYAML>=6.0.1
numpy>=1.26.4
ddgs>=0.1.0
cachetools>=5.3.0
redis>=5.0.0
# GLiNER2 for enhanced entity extraction in RAG
gliner2>=0.1.0

